+++ app/react/Thesauri/NewThesauri.js
@@ -1,9 +1,9 @@
 import React from 'react';
 
 import RouteHandler from 'app/App/RouteHandler';
-import ThesauriForm from 'app/Thesauris/components/ThesauriForm';
+import ThesauriForm from 'app/Thesauri/components/ThesauriForm';
 import { actions } from 'app/BasicReducer';
-import api from 'app/Thesauris/ThesaurisAPI';
+import api from 'app/Thesauri/ThesaurisAPI';
 
 export default class NewThesauri extends RouteHandler {
   static async requestState(requestParams) {
