+++ app/react/Thesauri/actions/thesauriActions.js
@@ -4,9 +4,9 @@ import superagent from 'superagent';
 import { actions as formActions } from 'react-redux-form';
 import { t } from 'app/I18N';
 import ID from 'shared/uniqueID';
-import * as types from 'app/Thesauris/actions/actionTypes';
+import * as types from 'app/Thesauri/actions/actionTypes';
 import { APIURL } from 'app/config';
-import api from 'app/Thesauris/ThesaurisAPI';
+import api from 'app/Thesauri/ThesaurisAPI';
 import * as notifications from 'app/Notifications/actions/notificationsActions';
 import { advancedSort } from 'app/utils/advancedSort';
 import { RequestParams } from 'app/utils/RequestParams';
