+++ app/react/Thesauri/components/specs/ThesauriFormGroup.spec.js
