+++ app/react/Thesauri/reducers/reducer.js
