+++ app/react/Thesauri/types/classifierModelType.d.ts
