+++ app/react/Thesauri/types/suggestionResultSchema.ts
