+++ app/react/Thesauri/types/suggestionResultType.d.ts
