+++ app/react/Thesauri/utils/suggestionQuery.ts
