+++ app/react/Thesauri/utils/valuesSort.ts
